test: admin.signOut() throws error when signed out
admin.signOut() throws `TypeError: Cannot read property 'session' of null` when a user is not signed in.  This commit contains documentation and failing unit/integration test updates in order to catch the condition where a user has attempted to sign out without being signed in and, rather than throwing the current error, throw an `UnauthenticatedError: Not signed in` error, instead.